# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Room Reservation",
        "description": "API Documentation for Hotel Room Reservation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/room": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Menampilkan semua kamar",
                "operationId": "1758cfa0a4f31164f32a8c2192f97d86",
                "responses": {
                    "200": {
                        "description": "List kamar berhasil diambil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Room"
                ],
                "summary": "Membuat kamar baru",
                "operationId": "9be38be1eed9b9d7e62fd45fca294b7c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "RoomNumber",
                                    "RoomType",
                                    "Capacity",
                                    "Status",
                                    "defaultPrice",
                                    "defaultExtraBedPrice"
                                ],
                                "properties": {
                                    "RoomNumber": {
                                        "type": "string"
                                    },
                                    "RoomType": {
                                        "type": "string",
                                        "enum": [
                                            "single",
                                            "double",
                                            "twin",
                                            "family",
                                            "suite"
                                        ]
                                    },
                                    "Capacity": {
                                        "type": "integer"
                                    },
                                    "Status": {
                                        "type": "string",
                                        "enum": [
                                            "ready",
                                            "maintenance"
                                        ]
                                    },
                                    "defaultPrice": {
                                        "type": "number"
                                    },
                                    "defaultExtraBedPrice": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kamar berhasil dibuat"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/room/{id}": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Menampilkan detail kamar berdasarkan ID",
                "operationId": "c45e99f9ccf62d45114462c46ea5036a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail kamar ditemukan"
                    },
                    "404": {
                        "description": "Kamar tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Room"
                ],
                "summary": "Update data kamar",
                "operationId": "096109bd729e9ee91d620d1b883d87df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "RoomNumber": {
                                        "type": "string"
                                    },
                                    "RoomType": {
                                        "type": "string",
                                        "enum": [
                                            "single",
                                            "double",
                                            "twin",
                                            "family",
                                            "suite"
                                        ]
                                    },
                                    "Capacity": {
                                        "type": "integer"
                                    },
                                    "Status": {
                                        "type": "string",
                                        "enum": [
                                            "ready",
                                            "maintenance"
                                        ]
                                    },
                                    "defaultPrice": {
                                        "type": "number"
                                    },
                                    "defaultExtraBedPrice": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kamar berhasil diupdate"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Room"
                ],
                "summary": "Hapus kamar berdasarkan ID",
                "operationId": "ba010fe020114b40d872de5e7913a8d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kamar berhasil dihapus"
                    },
                    "404": {
                        "description": "Kamar tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}